+++ NEWS.md
@@ -1,3 +1,7 @@
+# equate_2.1.0 (##########)
+
+* Added sample method for freqtab object, useful for bootstrapping.
+
 # equate_2.0.7 (2018-04-06)
 
 * Update to summary method for bootstrap output, to correct the calculation of mean standard error, bias, and root mean square error over score points.